Update dsassignment3part2.java
Write a program that repeatedly prompts the user to guess the capital of a randomly displayed state. Upon receiving the user input, the program reports whether the answer is correct. Your program must run a minimum of 5 times and display the total number of correct answers. The user’s answer is not case-sensitive.